server/requirements.txt
Flask
Flask-restful
Flask-SQLAlchemy
Flask-CORS
psycopg2-binary
passlib
python-dotenv
pillow
requests
waitress
PyJWT
